[refactor] 오타 수정 및 시리즈 화면 임시 페이지 전환
