+++ src/App.jsx
@@ -4,7 +4,7 @@ import { Routes, Route } from 'react-router-dom';
 
 import Header from '@components/Header';
 
-import HomePage from '@pages/HomPage';
+import HomePage from '@pages/HomePage';
 import NotFoundPage from '@pages/NotFoundPage';
 
 export default function App() {
@@ -14,6 +14,7 @@ export default function App() {
       <Divider />
       <Routes>
         <Route path="/" element={<HomePage />} />
+        <Route path="/series" element={<div>developing...</div>} />
         <Route path="*" element={<NotFoundPage />} />
       </Routes>
     </Container>
